feat(): update redis workspace detail refactor + optimze load
