+++ components/modules/redis-workspace/hooks/useRedisWorkspace.ts
@@ -196,8 +196,10 @@ export function useRedisWorkspace(options: {
     keys: browser.keys,
     databases: browser.databases,
     selectedKeyDetail: browser.selectedKeyDetail,
+    selectedKeyInfo: browser.selectedKeyInfo,
     loadingKeys: browser.loadingKeys,
     loadingSelectedKeyDetail: browser.loadingSelectedKeyDetail,
+    loadingSelectedKeyInfo: browser.loadingSelectedKeyInfo,
     savingValue: browser.savingValue,
     isDeletingKey: browser.isDeletingKey,
     selectedDatabaseIndex,
