+++ components/modules/redis-workspace/hooks/useRedisWorkspaceBrowser.ts
@@ -9,6 +9,7 @@ import type {
   RedisDatabaseOption,
   RedisDeleteResponse,
   RedisKeyDetail,
+  RedisKeyInfo,
   RedisKeyListItem,
   RedisValueUpdatePayload,
 } from '~/core/types/redis-workspace.types';
@@ -31,17 +32,21 @@ export function useRedisWorkspaceBrowser({
   const keys = ref<RedisKeyListItem[]>([]);
   const databases = ref<RedisDatabaseOption[]>([]);
   const selectedKeyDetail = shallowRef<RedisKeyDetail | null>(null);
+  const selectedKeyInfo = shallowRef<RedisKeyInfo | null>(null);
   const loadingKeys = ref(false);
   const loadingSelectedKeyDetail = ref(false);
+  const loadingSelectedKeyInfo = ref(false);
   const savingValue = ref(false);
   const editUnavailableReason = ref('');
   let selectedKeyDetailRequestId = 0;
+  let selectedKeyInfoRequestId = 0;
 
   const canEditSelectedValue = computed(
     () => !!selectedKeyDetail.value && !editUnavailableReason.value
   );
 
   const detailCache = new Map<string, RedisKeyDetail>();
+  const infoCache = new Map<string, RedisKeyInfo>();
 
   const getDetailCacheKey = (databaseIndex: number, key: string) =>
     `${connection.value?.id}:${databaseIndex}:${key}`;
@@ -92,14 +97,64 @@ export function useRedisWorkspaceBrowser({
     );
   };
 
+  const refreshSelectedKeyInfo = async (
+    key = session.value?.selectedKey,
+    options?: { force?: boolean }
+  ) => {
+    if (!connection.value || !session.value || !key) {
+      selectedKeyInfo.value = null;
+      loadingSelectedKeyInfo.value = false;
+      return;
+    }
+
+    const cacheKey = getDetailCacheKey(
+      session.value.selectedDatabaseIndex,
+      key
+    );
+    const cached = infoCache.get(cacheKey);
+
+    if (!options?.force && cached) {
+      selectedKeyInfo.value = cached;
+      loadingSelectedKeyInfo.value = false;
+      return;
+    }
+
+    const requestId = ++selectedKeyInfoRequestId;
+    loadingSelectedKeyInfo.value = true;
+
+    try {
+      const info = await $fetch<RedisKeyInfo>('/api/redis/browser/value/info', {
+        method: 'POST',
+        body: {
+          ...buildConnectionBody(connection.value),
+          databaseIndex: session.value.selectedDatabaseIndex,
+          key,
+        },
+      });
+
+      if (requestId !== selectedKeyInfoRequestId) {
+        return;
+      }
+
+      infoCache.set(cacheKey, info);
+      selectedKeyInfo.value = info;
+    } finally {
+      if (requestId === selectedKeyInfoRequestId) {
+        loadingSelectedKeyInfo.value = false;
+      }
+    }
+  };
+
   const refreshSelectedKeyDetail = async (
     key = session.value?.selectedKey,
     options?: { force?: boolean }
   ) => {
     if (!connection.value || !session.value || !key) {
       selectedKeyDetail.value = null;
+      selectedKeyInfo.value = null;
       editUnavailableReason.value = '';
       loadingSelectedKeyDetail.value = false;
+      loadingSelectedKeyInfo.value = false;
       return;
     }
 
@@ -111,10 +166,17 @@ export function useRedisWorkspaceBrowser({
 
     if (!options?.force && cached) {
       selectedKeyDetail.value = cached;
+      selectedKeyInfo.value = cached;
       editUnavailableReason.value = '';
       loadingSelectedKeyDetail.value = false;
+      loadingSelectedKeyInfo.value = false;
       return;
     }
+
+    // Fire the lightweight metadata-only call alongside the full value
+    // fetch so the header/badges can render before the (potentially slow)
+    // value read finishes.
+    void refreshSelectedKeyInfo(key, options);
 
     const requestId = ++selectedKeyDetailRequestId;
     selectedKeyDetail.value = null;
@@ -135,7 +197,9 @@ export function useRedisWorkspaceBrowser({
       }
 
       detailCache.set(cacheKey, detail);
+      infoCache.set(cacheKey, detail);
       selectedKeyDetail.value = detail;
+      selectedKeyInfo.value = detail;
       editUnavailableReason.value = '';
     } finally {
       if (requestId === selectedKeyDetailRequestId) {
@@ -189,14 +253,14 @@ export function useRedisWorkspaceBrowser({
         }
       );
 
-      detailCache.set(
-        getDetailCacheKey(
-          session.value.selectedDatabaseIndex,
-          session.value.selectedKey
-        ),
-        updatedDetail
+      const updatedCacheKey = getDetailCacheKey(
+        session.value.selectedDatabaseIndex,
+        session.value.selectedKey
       );
+      detailCache.set(updatedCacheKey, updatedDetail);
+      infoCache.set(updatedCacheKey, updatedDetail);
       selectedKeyDetail.value = updatedDetail;
+      selectedKeyInfo.value = updatedDetail;
       editUnavailableReason.value = '';
       toast.success('Redis key saved successfully', {
         description: `Updated ${session.value.selectedKey}`,
@@ -235,6 +299,7 @@ export function useRedisWorkspaceBrowser({
 
     store.patchSession(session.value.connectionId, { selectedKey: null });
     selectedKeyDetail.value = null;
+    selectedKeyInfo.value = null;
     await closeBrowserTabForConnection();
   };
 
@@ -255,9 +320,12 @@ export function useRedisWorkspaceBrowser({
         },
       });
 
-      detailCache.delete(
-        getDetailCacheKey(session.value.selectedDatabaseIndex, key)
+      const deletedCacheKey = getDetailCacheKey(
+        session.value.selectedDatabaseIndex,
+        key
       );
+      detailCache.delete(deletedCacheKey);
+      infoCache.delete(deletedCacheKey);
       await handleDeletedKeySelection([key]);
       await refreshKeys();
       toast.success('Redis key deleted', {
@@ -285,11 +353,14 @@ export function useRedisWorkspaceBrowser({
         },
       });
 
-      keysToDelete.forEach(key =>
-        detailCache.delete(
-          getDetailCacheKey(session.value!.selectedDatabaseIndex, key)
-        )
-      );
+      keysToDelete.forEach(key => {
+        const cacheKey = getDetailCacheKey(
+          session.value!.selectedDatabaseIndex,
+          key
+        );
+        detailCache.delete(cacheKey);
+        infoCache.delete(cacheKey);
+      });
       await handleDeletedKeySelection(keysToDelete);
       await refreshKeys();
       toast.success('Redis keys deleted', {
@@ -321,8 +392,10 @@ export function useRedisWorkspaceBrowser({
     keys,
     databases,
     selectedKeyDetail,
+    selectedKeyInfo,
     loadingKeys,
     loadingSelectedKeyDetail,
+    loadingSelectedKeyInfo,
     savingValue,
     isDeletingKey,
     editUnavailableReason,
@@ -330,6 +403,7 @@ export function useRedisWorkspaceBrowser({
     refreshKeys,
     refreshDatabases,
     refreshSelectedKeyDetail,
+    refreshSelectedKeyInfo,
     openKey,
     focusKey,
     saveSelectedValue,
