+++ core/data/changelogs/1.1.11.md
@@ -8,6 +8,8 @@ date: '2026-08-09'
 ## ✨ New Features
 
 - **Cloudflare D1 & Turso Shortcuts**: Cloudflare D1 and Turso now have their own cards in the Select Database Type step, so you can jump straight to their setup form instead of picking SQLite and digging into the Managed tab.
+- **Redis Key Delete**: Delete a key directly from the key detail panel, or select and delete one or multiple keys from the sidebar tree via a right-click context menu or the Delete/Backspace shortcut — all backed by a shared confirmation dialog.
+- **Redis Tree Context Menu**: Right-click a key or group in the Redis sidebar for quick delete actions.
 
 ## 🔧 Improvements
 
@@ -16,7 +18,15 @@ date: '2026-08-09'
   - SSL is no longer silently dropped when connecting via a Postgres connection string.
   - SSL certificate verification now follows the selected mode correctly (`require`/`preferred` no longer reject self-signed certs; `verify-ca`/`verify-full` enforce chain validation as expected).
   - SSH tunnels are more reliable under load: keepalive, reference-counted close, dead-tunnel eviction, and de-duplication of concurrent tunnel requests.
+- **Redis Instance Insights Overhaul**: Reworked Overview, Keyspace, Memory, Performance, Persistence, Replication, Clients, and Config tabs with new KPI cards and scope badges for clearer at-a-glance metrics.
+- **Faster Redis Key Browsing**: Key detail fetches now run in parallel and are cached per connection/database, so switching between keys — especially ones you've already opened — feels instant.
+- **Redis Tab Stability**: The Redis browser tab stays alive across tab switches, and the sidebar tree keeps focus correctly after clicks and deletes.
+- Deleting the key currently open in the browser now closes its tab automatically to keep the UI in sync.
 
 ## 🐛 Bug Fixes
 
 - Fixed Redis connection tests hanging indefinitely against an unreachable instance instead of reporting a failure.
+- Fixed Redis key scanning silently stopping at 500 results instead of scanning the full keyspace.
+- Fixed the delete confirmation dialog closing before the delete operation actually finished.
+- Fixed the bulk delete dialog taking a moment to open instead of showing instantly with a loading state.
+- Fixed opening connection details from the connection selector and status bar.
